Converted `asapdiscovery` into implicit namespace package
This repo features the core `asapdiscovery` Python package, but this
namespace structure allows for other packages, e.g. `asapdiscovery-foo`
to be created later with import semantics as `import asapdiscovery.foo`
that just work.

For comparison, this is exactly the approach taken by OpenFF for
`openff.toolkit`, `openff.evaluator`, etc., which appear under the same
`openff` namespace, but are developed in separate repos and also
packaged separately.